+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Toutes les modifications notables de ce projet seront documentées dans ce fichi
 - [#186](https://github.com/Refhi/Weda-Helper/issues/186) - "n" et "o" dans les fse même quand assurance "maternité" est sélectionnée
 - [#208](https://github.com/Refhi/Weda-Helper/issues/208) - Alt+W fonctionne depuis les pages d'accueil
 - [#209](https://github.com/Refhi/Weda-Helper/issues/209) - défaultPédia fonctionne désormais pour les ages <1 an
+- [#190](https://github.com/Refhi/Weda-Helper/issues/190) - accès facilité aux ATCD et aux notes patients depuis n'importe quelle liste de patients issus d'une recherche : Bouton de droite pour les notes, bouton du milieu pour les ATCD
 
 
 # [2.5.0] - 17/07/2024